Worked though a bunch of small bugs
Still need to debug more.

I would like to change some of the functionality. Mainly the update and
show sections, by adding options to find or update by a certain
criteria.

Attempted to style some, but quickly realized I should do this on a
different branch. I got caught in the weeds and broke some of the basic
functionality.

I hopefully repaired all that was broken and will work on a new branch
from here for styling.